refactor: renomeando header para melhor compreensão
