+++ src/components/patientHeader/PatientHeader.styles.ts
